+++ runepkg/gcc_command.sh
@@ -1,4 +1,0 @@
-
-
-# gcc command for testing...
-gcc -Wall -Wextra -std=c99 -D_GNU_SOURCE -o upkg upkg_cli.c upkg_config.c